==> Revisions/Complete_Workflow/39_creating_KML_ready_data.ipynb <==
import pandas as pd
import numpy as np

# Import predicted data sets
cities_df = pd.read_csv('/home/jeremy/tornadoesr/data/raw/cities_with_predictions.csv')
grid_df = pd.read_csv('/home/jeremy/tornadoesr/data/raw/grid_with_predictions.csv')

# Import the data set used for data transformations
unproc_tor_df = pd.read_csv("/home/jeremy/tornadoesr/data/raw/tor_data_inflation_adj.csv")
# Get the parameters for the transformations
# For lat
mean_lat = np.mean(unproc_tor_df['BEGIN_LAT'])
stand_dev_lat = np.std(unproc_tor_df['BEGIN_LAT'])
# For lon
mean_lon = np.mean(unproc_tor_df['BEGIN_LON'])
stand_dev_lon = np.std(unproc_tor_df['BEGIN_LON'])

# Undo the data transformations
cities_df['lon'] = cities_df['BEGIN_LON']*stand_dev_lon+mean_lon
cities_df['lat'] = cities_df['BEGIN_LAT']*stand_dev_lat+mean_lat
grid_df['lon'] = grid_df['BEGIN_LON']*stand_dev_lon+mean_lon
grid_df['lat'] = grid_df['BEGIN_LAT']*stand_dev_lat+mean_lat

# Get log10 damage for color coding and formatted property damage for display
# For the city predictions
cities_df['log10 property damage'] = np.log10(cities_df['DAMAGE_PROPERTY'] + 1)
cities_df['Property damage'] = ['$' + "{:,}".format(val) for val in cities_df['DAMAGE_PROPERTY']]
# For the grid predictions
grid_df['log10 property damage'] = np.log10(grid_df['DAMAGE_PROPERTY'] + 1)
grid_df['Property damage'] = ['$' + "{:,}".format(val) for val in grid_df['DAMAGE_PROPERTY']]

# Saving the data
cities_df[['lat', 'lon', 'Property damage', 'log10 property damage']].to_csv('testing1.csv', index = False)
grid_df[['lat', 'lon', 'Property damage', 'log10 property damage']].to_csv('testing2.csv', index = False)
